+++ cmd/create/cluster/flag.go
@@ -9,26 +9,26 @@ import (
 )
 
 const (
-	flagConfig     = "config"
-	flagKubeconfig = "kubeconfig"
-	flagOutput     = "output"
-	flagProvider   = "provider"
-	flagRelease    = "release"
+	flagConfig       = "config"
+	flagInstallation = "installation"
+	flagKubeconfig   = "kubeconfig"
+	flagOutput       = "output"
+	flagRelease      = "release"
 )
 
 type flag struct {
-	Config     string
-	Kubeconfig string
-	Provider   string
-	Output     string
-	Release    string
+	Config       string
+	Kubeconfig   string
+	Installation string
+	Output       string
+	Release      string
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVarP(&f.Config, flagConfig, "g", "", `The path to the file containing API endpoints and tokens for each provider.`)
 	cmd.Flags().StringVarP(&f.Kubeconfig, flagKubeconfig, "k", "", `The path to the directory containing the kubeconfigs for provider control planes.`)
 	cmd.Flags().StringVar(&f.Output, flagOutput, "", `The directory in which to store the cluster ID, kubeconfig, and provider of the created cluster.`)
-	cmd.Flags().StringVarP(&f.Provider, flagProvider, "p", "", `The provider of the target control plane.`)
+	cmd.Flags().StringVarP(&f.Installation, flagInstallation, "i", "", `The target management cluster type to be used ('aws', 'azure', 'kvm', or 'aws-china').`)
 	cmd.Flags().StringVarP(&f.Release, flagRelease, "r", "", `The semantic version of the release to be tested.`)
 }
 
@@ -44,8 +44,8 @@ func (f *flag) Validate() error {
 		if _, err := semver.NewVersion(f.Release); err != nil {
 			return microerror.Maskf(invalidFlagError, "--%s must be a valid semantic version", flagRelease)
 		}
-		if f.Provider == "" {
-			return microerror.Maskf(invalidFlagError, "--%s must be specified when defining an exact release version", flagRelease)
+		if f.Installation == "" {
+			return microerror.Maskf(invalidFlagError, "--%s must be specified when defining an exact release version", flagInstallation)
 		}
 	}