+++ cmd/create/release/flag.go
@@ -3,12 +3,15 @@ package release
 import (
 	"github.com/giantswarm/microerror"
 	"github.com/spf13/cobra"
+
+	"github.com/giantswarm/standup/pkg/key"
 )
 
 const (
 	flagConfig     = "config"
 	flagKubeconfig = "kubeconfig"
 	flagOutput     = "output"
+	flagPipeline   = "pipeline"
 	flagReleases   = "releases"
 )
 
@@ -16,6 +19,7 @@ type flag struct {
 	Config     string
 	Kubeconfig string
 	Output     string
+	Pipeline   string
 	Releases   string
 }
 
@@ -24,6 +28,7 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVarP(&f.Kubeconfig, flagKubeconfig, "k", "", `The path to the directory containing the kubeconfigs for provider control planes.`)
 	cmd.Flags().StringVar(&f.Output, flagOutput, "", `The directory in which to store the release name of the created release.`)
 	cmd.Flags().StringVarP(&f.Releases, flagReleases, "s", "", `The path of the releases repo on the local filesystem.`)
+	cmd.Flags().StringVarP(&f.Pipeline, flagPipeline, "t", key.DefaultPipelineName, `The name of the pipeline in which standup is currently running.`)
 }
 
 func (f *flag) Validate() error {